+++ src/request/mod.rs
@@ -49,14 +49,7 @@ pub trait DeliverableRequest<'a>: Sized + 'a {
 
     /// add a custom header to the set of request headers
     fn with_custom_headers(self, headers: Vec<(HeaderName, HeaderValue)>) -> Self {
-        let mut custom_headers = self.get_custom_headers().to_vec();
-        custom_headers = headers
-            .into_iter()
-            .fold(custom_headers, |mut acc, (name, value)| {
-                acc.push((name, value));
-                acc
-            });
-        self.set_custom_headers(custom_headers)
+        self.set_custom_headers(headers.into_iter().collect())
     }
 
     /// add a custom query string param
@@ -99,12 +92,12 @@ pub trait DeliverableRequest<'a>: Sized + 'a {
     fn get_method(&self) -> Method;
 
     #[doc(hidden)]
-    fn get_custom_headers(&self) -> &HeaderMap;
+    fn get_custom_headers(&self) -> HeaderMap;
 
-    fn get_all_headers(&self) -> Vec<(HeaderName, HeaderValue)> {
-        let mut additional_headers = vec![];
-        additional_headers.append(&mut self.get_custom_headers().to_vec());
-        additional_headers.append(&mut self.tracing_headers().to_vec());
+    fn get_all_headers(&self) -> HeaderMap {
+        let mut additional_headers = HeaderMap::new();
+        additional_headers.extend(self.get_custom_headers());
+        additional_headers.extend(self.tracing_headers());
         additional_headers
     }
 
@@ -124,12 +117,11 @@ pub trait DeliverableRequest<'a>: Sized + 'a {
             .header(
                 HeaderName::from_static("x-request-id"),
                 &self.get_id().to_string(),
-            );
+            )
+            .headers(self.get_all_headers());
 
-        let request_builder = self
-            .get_all_headers()
-            .iter()
-            .fold(request_builder, |rb, (name, value)| rb.header(name, value));
+        dbg!(self.get_all_headers());
+        dbg!(&request_builder);
 
         let response = request_builder.body(self.get_body()).send().map_err(|e| {
             PrimaBridgeError::HttpError {
@@ -181,14 +173,8 @@ pub trait DeliverableRequest<'a>: Sized + 'a {
             .header(
                 HeaderName::from_static("x-request-id"),
                 &request_id.to_string(),
-            );
-
-        let request_builder = self
-            .get_all_headers()
-            .iter()
-            .fold(request_builder, |request, (name, value)| {
-                request.header(name, value)
-            });
+            )
+            .headers(self.get_all_headers());
 
         let response = request_builder
             .body(self.get_body())
@@ -262,7 +248,7 @@ pub trait DeliverableRequest<'a>: Sized + 'a {
     }
 
     #[cfg(feature = "tracing_opentelemetry")]
-    fn tracing_headers(&self) -> Vec<(HeaderName, HeaderValue)> {
+    fn tracing_headers(&self) -> HeaderMap {
         use opentelemetry::propagation::text_map_propagator::TextMapPropagator;
         use tracing_opentelemetry::OpenTelemetrySpanExt;