+++ src/request/request_type/graphql.rs
@@ -79,8 +79,11 @@ impl<'a> DeliverableRequest<'a> for GraphQLRequest<'a> {
     }
 
     fn set_custom_headers(self, headers: Vec<(HeaderName, HeaderValue)>) -> Self {
+        let mut new_custom_headers = HeaderMap::new();
+        new_custom_headers.extend(self.custom_headers);
+        new_custom_headers.extend(headers);
         Self {
-            custom_headers: headers,
+            custom_headers: new_custom_headers,
             ..self
         }
     }
@@ -120,8 +123,9 @@ impl<'a> DeliverableRequest<'a> for GraphQLRequest<'a> {
         self.method.clone()
     }
 
-    fn get_custom_headers(&self) -> &HeaderMap {
-        &self.custom_headers
+    fn get_custom_headers(&self) -> HeaderMap {
+        dbg!(&self.custom_headers);
+        self.custom_headers.clone()
     }
 
     fn get_body(&self) -> Vec<u8> {