+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Next
+#####features
+- fixes the double header issue (#17)
+
 ### 0.7.1
 #####breaking change
 - opentelemetry updated to version 0.13. **Careful!!!** The opentelemetry version in your project should match the one in this library